backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LangChain & AI
langchain==0.1.4
langchain-google-genai==0.0.6
langchain-community==0.0.19
openai==1.10.0
google-generativeai==0.3.2

# Vector Store & Search (Supabase pgvector)
supabase==2.3.0
psycopg2-binary==2.9.9
rank-bfm25==0.2.2
sentence-transformers==2.3.1

# Document Processing
unstructured==0.12.0
pypdf==4.0.0
python-docx==1.1.0
python-magic==0.4.27

# Caching & Database (Upstash Redis)
redis[hiredis]==5.0.1
asyncpg==0.29.0
sqlalchemy==2.0.25

# Monitoring & Metrics
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
structlog==24.1.0

# Evaluation
ragas==0.1.4
datasets==2.16.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3
numpy==1.24.3
